+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Improved TCPConnector logging and only reconnect in client mode
 - physical-storage-api: allow VDO with lvm-thin
+- ETCD: Enabled DrbdOptions/auto-diskful-allow-cleanup on ctrl unless it was already set
 
 ### Fixed